Don't keep database open.
Keeping an open handle to the database might be good for performance but
it causes a problem terminating the server when the database is on an
NFS mount and the network is unavailable when trying to stop
calibre-web.

It is safer to only have the database open when it is being used. Then
network / NFS problems can only interfere at this time and it should be
possible to stop the server regardless of the availability of NFS mounts
and/or network at the time of shutdown.