Change CSS for title to avoid scrollbars
The margin on the title was causing vertical scrollbar. This lead to
resize events when switching tabs in Firefox, depending on the content
of the other tab (seems like a bug to me). Use padding instead of
vertical margin for a bit of space around the title and leave centering
to text-align.