Handle resize locally.
Firefox sometimes emits resize events when changing browser tabs,
depending on the content of the other tab. With variable size container,
epub.js listens for resize events and resizes everything when the window
is resized.

With a fixed size container, epub.js doesn't listen for resize events,
so they can be handled locally and the epub.js viewer resized only if
necessary.

For some reason, there are fewer resize events when the container size
is fixed. Changing tabs does not cause a resize event, regardless of the
contents of the other tab.